navbar component+some updates on dashboard
