+++ frontend/package.json
@@ -19,7 +19,8 @@
     "react-calendar": "^5.0.0",
     "react-csv": "^2.2.2",
     "react-dom": "^18.3.1",
-    "react-router-dom": "^6.26.2"
+    "react-router-dom": "^6.26.2",
+    "recharts": "^2.13.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.9.0",
