+++ frontend/src/components/ForDashboard/InfoSummary.jsx
@@ -1,0 +1,23 @@
+export default function InfoSummary() {
+  return (
+    <div className="info-summary">
+      <h3>Personal Information Summary</h3>
+      <p>
+        <strong>Name:</strong> Alan Walker
+      </p>
+      <p>
+        <strong>Job Title:</strong> Admin
+      </p>
+      <p>
+        <strong>Department:</strong> IT
+      </p>
+      <p>
+        <strong>Date of Birth:</strong> 05/05/1998
+      </p>
+      <p>
+        <strong>Working Since:</strong> 12/05/2020
+      </p>
+      <button className="btn-update">View Details</button>
+    </div>
+  );
+}
