+++ frontend/src/components/ForDashboard/QuickActions.jsx
@@ -1,0 +1,11 @@
+export default function QuickActions() {
+  return (
+    <div className="quick-actions">
+      <h3>Quick Actions</h3>
+      <p>Register New HR Manager</p>
+      <p>Register New Employee</p>
+      <p>Approve/Reject Leave Requests</p>
+      <p>Update Personal Information</p>
+    </div>
+  );
+}
