+++ frontend/src/components/ForDashboard/RemainingLeaves.css
@@ -1,0 +1,74 @@
+/* Card container for remaining leaves */
+.remaining-leaves-card {
+  padding: 20px;
+  background-color: #f9f9f9;
+  border-radius: 8px;
+  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+  text-align: center;
+  max-width: 700px;
+  margin: 0 auto;
+}
+
+/* Flexbox for the content: Leave cards and chart */
+.leave-content {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  flex-wrap: wrap;
+  margin-bottom: 20px;
+}
+
+/* Styling for individual leave cards */
+.leave-card {
+  flex: 1;
+  padding: 15px;
+  border: 1px solid #ddd;
+  border-radius: 8px;
+  background-color: #fff;
+  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+  font-size: 16px;
+  font-weight: bold;
+  color: #333;
+  margin: 10px;
+  text-align: center;
+  max-width: 150px;
+}
+
+.leave-card span {
+  display: block;
+  font-size: 24px;
+  color: #007bff;
+  margin-top: 10px;
+}
+
+/* Chart styling */
+.leave-chart {
+  flex: 1;
+  max-width: 300px;
+  margin: 10px auto;
+}
+
+/* Apply button styling */
+.btn-apply {
+  padding: 10px 20px;
+  background-color: #007bff;
+  color: #fff;
+  border: none;
+  border-radius: 8px;
+  cursor: pointer;
+}
+
+.btn-apply:hover {
+  background-color: #0056b3;
+}
+
+/* Responsive adjustments */
+@media (max-width: 768px) {
+  .leave-content {
+    flex-direction: column;
+  }
+
+  .leave-chart {
+    margin: 20px 0;
+  }
+}
