+++ frontend/src/components/ForDashboard/UserRoleDistribution.jsx
@@ -1,0 +1,11 @@
+export default function UserRoleDistribution() {
+  return (
+    <div className="user-role-distribution">
+      <h3>User Role Distribution</h3>
+      <p>Admin Users: 01</p>
+      <p>HR Managers: 05</p>
+      <p>Supervisors: 25</p>
+      <p>Regular Employees: 450</p>
+    </div>
+  );
+}
